test: assert too many request until cooldown
